+++ http/file_server.ts
@@ -517,22 +517,20 @@ function dirViewerTemplate(dirname: string, entries: EntryInfo[]): string {
       <body>
         <main>
           <h1>
-            Index of ${
-    headerPaths
-      .map((path, index) => {
-        if (path === "") return "";
-        const depth = headerPaths.length - index - 1;
-        let link;
-        if (depth == 0) {
-          link = ".";
-        } else {
-          link = "../".repeat(depth);
-        }
-        // deno-fmt-ignore
-        return html`<a href="${link}">${escape(path)}</a>`;
-      })
-      .join("/")
-  }/
+            Index of ${headerPaths
+              .map((path, index) => {
+                if (path === "") return "";
+                const depth = headerPaths.length - index - 1;
+                let link;
+                if (depth == 0) {
+                  link = ".";
+                } else {
+                  link = "../".repeat(depth);
+                }
+                // deno-fmt-ignore
+                return html`<a href="${link}">${escape(path)}</a>`;
+              })
+              .join("/")}/
           </h1>
           <table>
             <thead>
@@ -542,11 +540,10 @@ function dirViewerTemplate(dirname: string, entries: EntryInfo[]): string {
                 <th>Name</th>
               </tr>
             </thead>
-            ${
-    entries
-      .map(
-        (entry) =>
-          html`
+            ${entries
+              .map(
+                (entry) =>
+                  html`
                     <tr>
                       <td class="mode">
                         ${entry.mode}
@@ -559,9 +556,8 @@ function dirViewerTemplate(dirname: string, entries: EntryInfo[]): string {
                       </td>
                     </tr>
                   `,
-      )
-      .join("")
-  }
+              )
+              .join("")}
           </table>
         </main>
       </body>
