+++ net/deno.json
@@ -4,7 +4,7 @@
   "exports": {
     ".": "./mod.ts",
     "./get-available-port": "./get_available_port.ts",
-    "./ip-utils": "./ip_utils.ts",
+    "./ip": "./ip.ts",
     "./get-network-address": "./unstable_get_network_address.ts",
     "./unstable-get-network-address": "./unstable_get_network_address.ts"
   }
